Add a whitespace & semicolon expanded version of the 6k code.
Because this code was originally developed without version control,
and the shrinking process began long before it was feature complete,
there is stuff that is in the 6k version and not the main-branch
version.  If I recall correctly (after 12 years), some of that stuff
might be worthwhile.  With the original 6k code, it is impossible to
know; with this expanded version it is possibly possible.