Propagate castle state through tree search as integer (plus reformatting)
The castle state records whether and where castling is possible.  When
looking ahead for moves, the state (currently) remains unchanged, so
it can get out of sync with reality, allowing bizarre castlings to be
attempted that can leave the board in messy state (to be precise, the
rooks multiply).

To keep the castling state up to date, it needs to be altered while
climbing up and down the tree.  When the state is stored as a mutable
list, the changes need to be recorded and made both ways; with
immutable bit flags, the changes propagate one way and revert
automatically. So this change makes things simpler, despite
appearances.

The state is not actually changed yet. Just propagated.