Increase pawn centre weighting, and use idiomatic loop construct
The pawn weighting used to have a 0.5 multiplier, which some time ago
was affected by the introduction of a parseInt(), rounding down all
the values, and the pawns often took just one step.

The old loop went for(z=99;z>20;z--), which saved one byte relative to
for(i=21;i<100;i++).