Slightly diminish value of distant inevitable checkmate
When treeclimber is used to find checkmate, it can't tell a series of
forced moves resulting in checkmate from the immediate thing.  If those
forced moves push some other buttons along the way (e.g. capture material),
they will be preferred.  And to cut a long story short and brush over
some lack of understanding, the real checkmate is missed.  By
subtracting some of the power of the checkmate at each level, the quickest
one is preferred.


Also, increase value of king, for a safety margin.