Use adaptive search depth at higher levels of play.
So far this is very simple.  If the search doesn't take very long at
all, it will be retried at the next depth.  And again, if it still
doesn't take long.

This is slightly wasteful, in that the original search is thrown away,
and the level is recalibrated every move, not estimated based on
previous moves.  But this waste is not huge, given that each deeper
search is almost always an order of magnitude slower than the
shallower one, so the waste is in the order of a few percent.