Don't mistake a knight for a rook when checking for castling
The sliding pieces -- bishops, rooks, queens -- have the 2 bit clear
and the 4 and 8 bits set for diagonal and straight moves
respectively.

When checking for castling blockers, the 2 bit wasn't checked, so the
knight could be mistaken for a rook and a king for a queen.

Again based on a report from Ivan Yelizariev.  Thank you!