Simplify treeclimber to plain alpha-beta search
This *seems* to work properly (and is much faster) but it is hard to know
without proper tests.

The benefit presumably comes from dropping the sort, but a few other
changes have been made, like the test for an empty pseudo-legal move
list.  In theory, an empty move list will happen if every piece
is physically blocked from moving.