Reintroduce negascout without move sorting
This is faster (25% less time) and ought to (and seems to) produce
equivalent results.

On chromium 17, this is takes ~192ms for a test that took ~260ms
with the alpha-beta, and ~960ms with the best of the old sorted
PV/negascout search.

On Firefox 11, the numbers are 308 vs 420 and 2450, respectively.

Epiphany 3.0.4 dropped from 4067 to 651 overall.

Thus this new version is between 5 and 8 times faster.  The selected
moves are not always the same as the old version, which is to be
expected because the moves evaluated in a different order, and scoring
ties are resolved differently.  But it is the same as the plain alpha-beta
of the previous commit.