Fix github issue #5 -- missed queen or rook on king side
To check for queens and rooks on the king's home rank in the direction
opposite the castling rook, we slide along from the given starting
position.  But in the queen side case, the starting position is
actually the king's ending position, which simplifies some stuff but
not this because the sliding search always hits the king before the
queen or rook.