The true value of a promoted pawn is (queen - pawn), regardless
If you choose a knight, you must have a really good reason, which
makes it more valuable than a queen, actually.

And if a player selects a knight as the next promotion, it can distort
the search, because they can switch to queen at the last minute. So
assuming queen is right short of something much cleverer that parses
all the alternatives.