Shift pawn promotion state into display.js (prompted by Chanibal)
In github issue #7, Chanibal pointed out that the pawn promotion code
is a mess, and corrected some of the problems by always setting the
promotion state right before making the move.

This made it obvious that storing the promotion state does nothing
useful.  But it actually it does subtle harm, because the same
promotion is used for all levels of search, rather than just for the
immediate pawn move.

This commit makes the chosen promotion a parameter of p4_move(), with
an undefined promotion assumed to a queen.  This gets passed to
treeclimber, which makes the same assumption about undefined values.
Treeclimber drops the promotion when recursing -- it does not assume
future promotions will be the same as this one.

A potential downside is the cost of adding an 11th parameter to
treeclimber(), but that doesn't appear to of measurable size.