Try not to mistake stalemate for checkmate
When every possible move results in the king being taken, treeclimber
will give that node a very losing score.  But that gives the same
score to stalemate as checkmate.  Now when all seems lost, treeclimber
tries a null move to see whether the king is taken, and this score is
used.

This is effectively a check test, and an inefficient one at that.
Which would matter if it happened more often.

A big flaw is it doesn't distinguish between immediate loss of the
king, and deeper inevitable loss, where the check test is irrelevant.