convert p4_move flags to a move result object, with move string
The thinking behind this is that a proper short algebraic move description
(as in "Rb5") occasionally needs a qualifier to distinguish which piece is
being moved (e.g., "ReB5" or "R3b5").  It would be simple to always add
them, but that is bad form -- the correct thing is to add them only as
necessary.  To do that you really want knowledge of the board state
*before* the move, so it makes sense for that string to be generated during
the call to p4_move, and not later.  Commit fc6021047544 added the necessary
routine, and this tries to hook it up without causing too much damage.

NB: the strings it generates are sometimes over-qualified, in that it is
based on a list of pseudo-legal moves. It would not strictly be necessary
to say "ReB5" if the alternative "R3b5" would put you in check.

Eventually the flags could be dropped altogether.