Add a random number generator (Jenkins' small one, modified for JS)
This is for Zobrist hashing for transposition tables, for shuffling
moves, and giving weight jitter.

The 48 bit function is for additive Zobrist hashing (32 bits are too
small, and 64 bits is too big for javascript's double floating point
numbers, and xor would truncate to 32 bits so addition is necessary).

The 31 bit function works around javascript's tendency to convert an
unsigned 32 bit number with high bit set into a negative number.