Use RNG to shuffle pieces; add jitter to later weights
The pieces list is generated in board order which means that tied
scores favoured pieces in one corner of the board.  Now the list is
shuffled in p4_prepare(), so each move uses a different tie
resolution order for its search.

Also, in the endgame, the winning side adds a bit of random jitter to
its weights, because it wants to avoid a draw through repetition. This
possibly doesn't work entirely: repetition is lessened, but the
changes affect the reasoning of the losing side as much as the winning
one, so they end up avoiding a draw they should want.