Calculate en passant as needed from pawn delta, with many consequences
Rather than flag en passant in p4_parse(), then pass that flag around
all over the place, it is simpler to decide it by considering whether
a pawn move being considered in treeclimber covers two rows.

A consequence of that is that state.enpassant can't be passed in to
treeclimber from the top (unless some magical fake pawn move is used, or
the previous move is unwound and replayed).  One way around that is to
have a special top-level that uses state.enpassant, and because a special
top level is good in other ways, and is quite simple, that is what I did.
Now p4_findmove() is that top level (whereas it used to be a thin
treeclimber wrapper). One advantage of having a special root node is that
the best moves don't need to be fed back through treeclimber, just the
scores, thus further reducing pointless array manipulation.