Weights value passed pawns a little
If a pawn has no opposing pawn ahead of it, it is easier for it to get
to the other end. Obviously pawns prefer this.

The trouble with calculating this kind of thing in p4_prepare() is
that the board will change during the tree search and the weights
will be obsolete.