Parse scores removal of weight of captured piece
Now when a piece is captured, as well as its plain value being
removed, its current weighting is subtracted from the score.  This
seems the right thing to do, but is little tested.

Previously (as a gross example), a pawn could be weighted so heavily
to move into the centre early that its immediate capture might seem
like an even exchange.