Allow display.js to run multiple boards with simpler API
So there's this P4wn_display object that encapsulates all the I/O
along with the p4wn engine state, and a p4wnify() function that
converts an html element into a chess board.

This is work in progress, and may not actually work with multiple
boards (though it dies work with one). Also breaks fen-test.js for
now.