top-level treeclimber is an atttribute of state
So when you switch the search method it changes the treeclimber
attribute and doesn't re-assign p4_treeclimber.

This means different boards on the same page can have their own
treeclimbers, which is pretty useless in itself.  But in the future
different search methods will be able to be set upon each other for
testing and automatic tuning, and that will be good.