p4_parse search piece list backwards, putting recently moved first
Moved pieces get appended on the end of the pieces list.  Alpha-beta
searches are faster if the best moves are tried early on, and the best
move is (by this hypothesis) likely to involve a recently moved piece.

Approximately 2-fold speed increase on the benchmark board, but less
on others.  At the beginning the hypothesis is clearly wrong, where
good play involves getting all your pieces out, not bouncing one piece
around.