Reorder pieces to reorder movelists
The move list is constructed by finding the moves of each piece in
(reverse) order.  If the pieces list is well arranged, the best moves
will be found first, making alpha-beta derived searches much faster.

This first iteration orders the pieces by their best weighted moves --
that is it puts capturers first, in the order of the pieces they
capture, then other pieces in the order of the weight of their moves.

Captures already go first, of course; this promotes the other moves of
the catpuriong piece.

It does nothing to the default benchmark, but doubles the speed of the
initial position -- presumably because move quality there is all about
positional weights.