Prune out duplicate moves in counting treeclimber
p4_counting_treeclimber is used to test the move generation code
against known move counts (known as "perft testing" for some reason).

It was finding extra moves because state.pieces lists can contain
duplicates if a piece moves back to a place it used to be.

These duplicates are *probably* rare enough to be harmless otherwise.