Modify pieces list in place
Instead of lengthening and checking on the fly in p4_parse.

This shifts a check from p4_parse to p4_make_move, which means it
happens ~30 times less often. OTOH, the remedy is far more expensive
(list munging vs skipping).

Also without this the damage to the pieces list is cumulative, so with
deeper searches the p4_parse check will fail more often.