Recognise draws due to insufficient material
With certain combinations of pieces in the end-game, a win is
impossible however badly the would-be loser plays. These would end up
being draws under the 50-move rule, but it is tedious to wait for
that, and the rules of chess allow short-circuit draws to be claimed.

Thanks to Greg Mayer for reporting this.
http://sourceforge.net/mailarchive/message.php?msg_id=30657626

Reported-by: Greg Mayer <greg@fingerfoodstudios.com>