Fix FEN import with impossible castling claims
It looks as if github issues #16, #17, and #18 have been caused by a
FEN string that claims castling is possible when it clearly is not
(the king is not at home).

Here we add checks to FEN import to ignore these impossible claims,
and add a few tests.

Thanks to Adrian Dușa for the reports.