tests: hack to run under nodejs
To run things properly under node, the tests would need to use the
web-incompatible commonjs module system, and to make it work both as
web and node modules would require annoying invasive changes, with
possibly subtle incompatibilities.

By concatenating the scripts together, we run it on node as it is run
in the browser. The only differences are we include "auto-test-node.js"
instead of "auto-test-html.js", and we run under "strict" mode. Strict
mode would work in the browser too, but might interact badly with
other scripts on the page.